Don't troncate to int a long
git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1687932 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 7a9d006a2bb46090a872b1916ab608dfabb754d3

+++ src/components/org/apache/jmeter/sampler/TestAction.java
@@ -109,9 +109,9 @@ public class TestAction extends AbstractSampler implements Interruptible {
     }
 
     private void pause(String mili_s) {
-        int milis;
+        long milis;
         try {
-            milis=Integer.parseInt(mili_s);
+            milis=Long.parseLong(mili_s);
         } catch (NumberFormatException e){
             log.warn("Could not create number from "+mili_s);
             milis=0;
